+++ dst/app/routes/orders.js
@@ -24,6 +24,7 @@ const rateLimit_1 = require("../middlewares/rateLimit");
 const validator_1 = require("../middlewares/validator");
 const iam_1 = require("../iam");
 const redis = require("../../redis");
+const connectMongo_1 = require("../../connectMongo");
 const ADDITIONAL_PROPERTY_VALUE_MAX_LENGTH = (process.env.ADDITIONAL_PROPERTY_VALUE_MAX_LENGTH !== undefined)
     ? Number(process.env.ADDITIONAL_PROPERTY_VALUE_MAX_LENGTH)
     // tslint:disable-next-line:no-magic-numbers
@@ -292,8 +293,11 @@ ordersRouter.get('/download', permitScopes_1.default([iam_1.Permission.User, 'or
         .isISO8601()
         .toDate()
 ], validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+    let connection;
     try {
-        const orderRepo = new cinerino.repository.Order(mongoose.connection);
+        // 長時間占有する可能性があるのでコネクションを独自に生成
+        connection = yield connectMongo_1.connectMongo({ defaultConnection: false });
+        const orderRepo = new cinerino.repository.Order(connection);
         const searchConditions = Object.assign(Object.assign({}, req.query), { project: { ids: [req.project.id] } });
         const format = req.query.format;
         const stream = yield cinerino.service.report.order.stream({
@@ -301,9 +305,22 @@ ordersRouter.get('/download', permitScopes_1.default([iam_1.Permission.User, 'or
             format: format
         })({ order: orderRepo });
         res.type(`${req.query.format}; charset=utf-8`);
-        stream.pipe(res);
+        stream.pipe(res)
+            .on('error', () => __awaiter(void 0, void 0, void 0, function* () {
+            if (connection !== undefined) {
+                yield connection.close();
+            }
+        }))
+            .on('finish', () => __awaiter(void 0, void 0, void 0, function* () {
+            if (connection !== undefined) {
+                yield connection.close();
+            }
+        }));
     }
     catch (error) {
+        if (connection !== undefined) {
+            yield connection.close();
+        }
         next(error);
     }
 }));