+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - IAMロールリポジトリを追加
 - アプリケーションクライアントもIAMメンバーとして管理するように調整
+- Mongoのcursorを使用するアクションに関して、DBコネクションを独自に生成するように変更
 
 ### Deprecated